格式修改 Demo类sum字段由Integer类型改为Double类型

+++ src/main/java/com/zhangke/gitdemo/entity/Demo.java
@@ -7,13 +7,13 @@ package com.zhangke.gitdemo.entity;
  */
 public class Demo {
 
-    Integer sum;
+    Double sum;
 
-    public Integer getSum() {
+    public Double getSum() {
         return sum;
     }
 
-    public void setSum(Integer sum) {
+    public void setSum(Double sum) {
         this.sum = sum;
     }
 }
